+++ src/main/java/com/lazerycode/jmeter/results/IResultScanner.java
@@ -1,9 +1,11 @@
-package com.lazerycode.jmeter.testrunner;
+package com.lazerycode.jmeter.results;
 
 public interface IResultScanner {
 
     int getSuccessCount();
 
     int getFailureCount();
+
+    int getTotalCount();
 
 }