+++ src/main/java/com/lazerycode/jmeter/testrunner/TestFailureDecider.java
@@ -1,33 +1,33 @@
 package com.lazerycode.jmeter.testrunner;
 
+import com.lazerycode.jmeter.results.IResultScanner;
+
 /**
  * Since 3.8.1
- *
  */
 public class TestFailureDecider {
 
-    private IResultScanner resultScanner;
-    private boolean ignoreResultFailures;
-    private float errorPercentageThreshold;
-    private float errorPercentage;
+    private final IResultScanner resultScanner;
+    private final boolean ignoreResultFailures;
+    private final double errorPercentageThreshold;
+    private double errorPercentage;
     private boolean checkRan;
-    
-    public TestFailureDecider(boolean ignoreResultFailures, float errorPercentageThreshold, IResultScanner resultScanner) {
+
+    public TestFailureDecider(boolean ignoreResultFailures, double errorPercentageThreshold, IResultScanner resultScanner) {
         this.ignoreResultFailures = ignoreResultFailures;
         this.errorPercentageThreshold = errorPercentageThreshold;
         this.resultScanner = resultScanner;
     }
 
     public boolean failBuild() {
-        if(!checkRan) {
+        if (!checkRan) {
             throw new IllegalStateException("You need to call runChecks");
         }
-        return !ignoreResultFailures &&  errorPercentage > errorPercentageThreshold;
+        return !ignoreResultFailures && errorPercentage > errorPercentageThreshold;
     }
 
     public void runChecks() {
-        this.errorPercentage = (float)resultScanner.getFailureCount() / 
-                (float)(resultScanner.getSuccessCount() + resultScanner.getFailureCount()) * 100;
+        this.errorPercentage = (double) Math.round(resultScanner.getFailureCount() * 10000.0 / resultScanner.getTotalCount()) / 100;
         this.checkRan = true;
     }
 
@@ -41,15 +41,15 @@ public class TestFailureDecider {
     /**
      * @return the errorPercentageThreshold
      */
-    public float getErrorPercentageThreshold() {
+    public double getErrorPercentageThreshold() {
         return errorPercentageThreshold;
     }
 
     /**
      * @return the errorPercentage
      */
-    public float getErrorPercentage() {
+    public double getErrorPercentage() {
         return errorPercentage;
     }
-    
+
 }